requirements.txt
discord.py==2.4.0
google-genai
pytubefix
beautifulsoup4
python-weather
wolframalpha
nest_asyncio
wikipedia
protobuf
nest-asyncio
httpx
xmltodict
requests
duckduckgo-search
matplotlib
numpy
samplerate
discord-ext-voice-recv
pytest
regex
pytest-mock
docker
tqdm
anyio